refactor(domain): move entidades para seus respectivos domínios
