+++ src/main/java/com/project/miinhareceita/config/customgrant/CustomUserAuthorities.java
@@ -1,8 +1,8 @@
 package com.project.miinhareceita.config.customgrant;
 
-import java.util.Collection;
-
 import org.springframework.security.core.GrantedAuthority;
+
+import java.util.Collection;
 
 public class CustomUserAuthorities {
