+++ src/main/java/com/project/miinhareceita/controllers/FavoriteController.java
@@ -2,7 +2,6 @@ package com.project.miinhareceita.controllers;
 
 import com.project.miinhareceita.dtos.FavoriteDTO;
 import com.project.miinhareceita.dtos.FavoriteInsertDTO;
-import com.project.miinhareceita.dtos.RecipeDTO;
 import com.project.miinhareceita.services.FavoriteService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,9 +9,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.net.URI;
 import java.util.List;
 
 @RestController
