+++ src/main/java/com/project/miinhareceita/dtos/RecipeFavoriteDTO.java
@@ -1,7 +1,7 @@
 package com.project.miinhareceita.dtos;
 
 
-import com.project.miinhareceita.entities.Recipe;
+import com.project.miinhareceita.recipe.domain.Recipe;
 
 import java.time.Instant;
