+++ src/main/java/com/project/miinhareceita/recipe/domain/Recipe.java
@@ -1,8 +1,11 @@
-package com.project.miinhareceita.entities;
+package com.project.miinhareceita.recipe.domain;
 
 
+import com.project.miinhareceita.category.domain.Category;
+import com.project.miinhareceita.favorite.domain.Favorite;
+import com.project.miinhareceita.review.domain.Review;
+import com.project.miinhareceita.user.domain.User;
 import jakarta.persistence.*;
-import lombok.*;
 
 import java.time.Instant;
 import java.util.*;
