+++ src/main/java/com/project/miinhareceita/repositories/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.project.miinhareceita.repositories;
 
-import com.project.miinhareceita.entities.Category;
+import com.project.miinhareceita.category.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
