+++ src/main/java/com/project/miinhareceita/repositories/RecipeIngredientsRepository.java
@@ -1,6 +1,6 @@
 package com.project.miinhareceita.repositories;
 
-import com.project.miinhareceita.entities.RecipeIngredients;
+import com.project.miinhareceita.recipe.domain.RecipeIngredients;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
