+++ src/main/java/com/project/miinhareceita/repositories/RecipeRepository.java
@@ -1,16 +1,14 @@
 package com.project.miinhareceita.repositories;
 
 import com.project.miinhareceita.dtos.RecipeMinDTO;
-import com.project.miinhareceita.entities.Recipe;
 import com.project.miinhareceita.projections.RecipeProjections;
-import com.project.miinhareceita.projections.RecipeWithAverageProjection;
+import com.project.miinhareceita.recipe.domain.Recipe;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.time.Instant;
 import java.util.List;
 
 @Repository
