+++ src/main/java/com/project/miinhareceita/repositories/ReviewRepository.java
@@ -1,9 +1,7 @@
 package com.project.miinhareceita.repositories;
 
-import com.project.miinhareceita.entities.Review;
 import com.project.miinhareceita.projections.ReviewProjections;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import com.project.miinhareceita.review.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
