+++ src/main/java/com/project/miinhareceita/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
 package com.project.miinhareceita.repositories;
 
-import com.project.miinhareceita.entities.Role;
+import com.project.miinhareceita.user.domain.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
