+++ src/main/java/com/project/miinhareceita/services/validation/UserInsertValid.java
@@ -1,12 +1,12 @@
 package com.project.miinhareceita.services.validation;
 
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
 
 @Constraint(validatedBy = UserInsertValidator.class)
 @Target({ ElementType.TYPE })
