+++ src/main/java/com/project/miinhareceita/user/domain/Role.java
@@ -1,12 +1,7 @@
-package com.project.miinhareceita.entities;
+package com.project.miinhareceita.user.domain;
 
+import jakarta.persistence.*;
 import lombok.*;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import org.springframework.security.core.GrantedAuthority;
 
 @AllArgsConstructor
