+++ src/test/java/com/project/miinhareceita/services/UserServiceTest.java
@@ -2,8 +2,6 @@ package com.project.miinhareceita.services;
 
 import com.project.miinhareceita.dtos.UserDTO;
 import com.project.miinhareceita.dtos.UserInsertDTO;
-import com.project.miinhareceita.entities.Role;
-import com.project.miinhareceita.entities.User;
 import com.project.miinhareceita.projections.UserDetailsProjection;
 import com.project.miinhareceita.repositories.RoleRepository;
 import com.project.miinhareceita.repositories.UserRepository;
